@@ -3,27 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.6.0",
-    "@emotion/styled": "^11.6.0",
-    "@material-ui/core": "^5.0.0-beta.5",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "ali-oss": "^6.16.0",
-    "axios": "^0.24.0",
-    "react": "^17.0.2",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-dom": "^17.0.2",
-    "react-pageflip": "^2.0.3",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "react-select": "^5.4.0",
-    "web-vitals": "^1.1.2",
-      "antd": "^4.13.0",
-    "bootstrap": "^5.1",
-    "reactstrap": "^8.10.0"
-
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.9",
+    "@mui/material": "^5.10.9",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.5.0",
+    "web-vitals": "^3.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,8 +41,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "react-app-rewired": "^2.1.8"
   }
 }
