@@ -1,16 +1,16 @@
-import React from 'react'
-import { Layout } from "antd";
-import FeedingList from '../../components/FeedingList'
-import "./main.css"
-
-const { Footer, Content } = Layout;
-
-export default function Main(props) {
-    return (
-        <Layout style={{ height: "100vh", backgroundColor: "white" }}>
-            <Content style={{ alignSelf: "center" }}>
-                <FeedingList history={ props.history } />
-            </Content>
-        </Layout>
-    )
-}
+// import React from 'react'
+// import { Layout } from "antd";
+// import FeedingList from '../../components/FeedingList'
+// import "./main.css"
+//
+// const { Footer, Content } = Layout;
+//
+// export default function Main(props) {
+//     return (
+//         <Layout style={{ height: "100vh", backgroundColor: "white" }}>
+//             <Content style={{ alignSelf: "center" }}>
+//                 <FeedingList history={ props.history } />
+//             </Content>
+//         </Layout>
+//     )
+// }
