@@ -1,8 +1,6 @@
 import React, { Component } from 'react';
 import '../../App.css';
-import AppNavbar from '../../components/AppNavbar';
 import { Link } from 'react-router-dom';
-import { Button, Container } from 'reactstrap';
 
 
 
@@ -13,25 +11,21 @@ class Home extends Component {
 
         return (
             <div>
-                <AppNavbar/>
-                <Container fluid>
-                    {/* <Button color="link"><Link to="/feeding">feeding</Link></Button> */}
-                    <Button color="link"><Link to="/login">login</Link></Button>
-                    <Button color="link"><Link to="/register">register</Link></Button>
-                    <Button color="link"><Link to="/feeding">feeding</Link></Button>
-                    <Button color="link" onClick={() => { this.logout() }}>logout</Button>
+                <nav>
+                    <Link to="/">home</Link>
+                </nav>
+                <nav>
+                    <Link to="/home2">home2</Link>
+                </nav>
+                <nav>
+                    <Link to="/login">login</Link>
+                </nav>
+                <nav>
+                    <Link to="/register">register</Link>
+                </nav>
 
-                </Container>
             </div>
         );
-
-
-    }
-
-
-    logout =()=> {
-        console.log(11);
-        localStorage.removeItem("token");
     }
 }
 
