@@ -1,16 +1,16 @@
-import React from 'react'
-import { Result, Button } from "antd";
-import { Link } from "react-router-dom";
-import "./not-found.css";
-export default function NotFound() {
-    return (
-        <div className="not-found-page">
-            <Result 
-                status="404"
-                title="404"
-                subTitle="nothing here ..."
-                extra={ <Button type="primary"><Link to="/">back to home</Link></Button> }
-            />
-        </div>
-    )
-}
+// import React from 'react'
+// import { Result, Button } from "antd";
+// import { Link } from "react-router-dom";
+// import "./not-found.css";
+// export default function NotFound() {
+//     return (
+//         <div className="not-found-page">
+//             <Result
+//                 status="404"
+//                 title="404"
+//                 subTitle="nothing here ..."
+//                 extra={ <Button type="primary"><Link to="/">back to home</Link></Button> }
+//             />
+//         </div>
+//     )
+// }
